API call change
First idea was to query the liked videos then see which are from the particular channel, but the process didn't seem efficient. Since each API call is only up to 50 results and I have a Liked Videos playlist of >5000, the call would be computationally expensive. Think a storage system to immediately query and store those results for future reference would be the way to go.

My idea to get started now is to query all the videos on the page with a new API call and add some kind of indication that that video was liked. The process will rerun as the user scrolls down the page and more videos (thumbnails) load in.